requirements.txt
openai>=1.0.0
flask==3.0.3
python-dotenv==1.0.1
requests==2.31.0
beautifulsoup4==4.12.2
reportlab==4.0.4
sqlalchemy==2.0.23
flask-sqlalchemy==3.1.1
uuid4==1.1.4
datetime
json5==0.9.14